Added in Meloncholic Skull in small rocky area near spawn.
If the worldgen is not what we desire, we can direct edit the spawning setpiece to force it to spawn by the portal. We MUST do a direct edit to do this though to my knowlege. And I was told not to code like that when I joined the team.